Refactor ApplicationFlow to be cleaner and fit the new VH app flow
