--- apps/site/src/lib/components/forms/shared/ApplyConfirmation/ConfirmationDetails.module.scss
@@ -1,16 +1,0 @@
-.header {
-	color: var(--color-brown);
-}
-
-.details {
-	background-color: #ffffff;
-	border-radius: 10px;
-
-	& > p {
-		color: #432810;
-	}
-}
-
-.policy {
-	color: var(--color-brown);
-}
