+++ apps/site/src/lib/utils/useForm.ts
@@ -1,9 +1,7 @@
 "use client";
 
-import { FormEvent, PropsWithChildren, useState } from "react";
+import { FormEvent, useState } from "react";
 import axios from "axios";
-
-import Button from "@/lib/components/Button/Button";
 
 import hasDeadlinePassed from "@/lib/utils/hasDeadlinePassed";
 
@@ -15,16 +13,7 @@ const FIELDS_WITH_OTHER = [
 	"experienced_technologies",
 ];
 
-interface BaseFormProps {
-	applicationType: "Hacker" | "Mentor" | "Volunteer";
-	applyPath: string;
-}
-
-export default function BaseForm({
-	applicationType,
-	applyPath,
-	children,
-}: PropsWithChildren<BaseFormProps>) {
+export default function useForm(applyPath: string) {
 	const [submitting, setSubmitting] = useState(false);
 	const [sessionExpired, setSessionExpired] = useState(false);
 
@@ -88,39 +77,9 @@ export default function BaseForm({
 		setSubmitting(false);
 	};
 
-	const sessionExpiredMessage = (
-		<p className="text-red-500 w-11/12">
-			Your session has expired. Please{" "}
-			<a href="/login" target="_blank" className="text-blue-600 underline">
-				log in from a new tab
-			</a>{" "}
-			to restore your session and then try submitting again.
-		</p>
-	);
-
-	return (
-		<form
-			method="post"
-			className="bg-black border-[5px] border-white text-[var(--color-white)] w-8/12 flex flex-col items-center py-12 gap-14 z-1 max-[800px]:w-9/12 max-[400px]:w-11/12 drop-shadow-[25px_33px_0px_rgba(255,255,255,1)]"
-			action={applyPath}
-			encType="multipart/form-data"
-			onSubmit={handleSubmit}
-		>
-			<input
-				type="text"
-				name="application_type"
-				value={applicationType}
-				readOnly
-				hidden
-			/>
-			{children}
-			<Button
-				text="Submit"
-				className="text-2xl !px-11 !py-2"
-				isLightVersion={true}
-				disabled={submitting}
-			/>
-			{sessionExpired && sessionExpiredMessage}
-		</form>
-	);
+	return {
+		submitting,
+		sessionExpired,
+		handleSubmit,
+	};
 }
